SUKU fetch config fixed, heartbeat moved from Serial to Runtime
